Rewrite color picker
